+++ packages/file-storage/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 This is the changelog for [`file-storage`](https://github.com/remix-run/remix/tree/main/packages/file-storage). It follows [semantic versioning](https://semver.org/).
 
-## Unreleased
+## v0.11.0 (2025-11-05)
 
 - Move `@remix-run/lazy-file` to `peerDependencies`
 - Build using `tsc` instead of `esbuild`. This means modules in the `dist` directory now mirror the layout of modules in the `src` directory.